fix extent and zoom on report under certain conditions, update canopyDensity on report "
